requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
psycopg[binary]==3.1.18
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.26.4
pandas==2.2.1
scipy==1.12.0
python-dateutil==2.8.2
pytest==8.1.1
httpx==0.27.0
matplotlib==3.8.3
